requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain-openai
langchain-pinecone
unstructured[pdf]
pinecone-client
openai
tiktoken